content_repurpose_agent/requirements.txt
streamlit
google-generativeai>=0.7.0
nltk
sentence-transformers
PyPDF2
python-docx
python-dotenv
sqlite-utils
fastapi[standard]

